refac: mv patch specs to own module
